Did a cleanup. #59
Removed all openstack related stuff

Moved playbooks that install a single role that is also installed via
the cluster.yml. This keeps the root of the repository less of a mess in
my oppinion.

As famous quote goes "There are only two hard things in
Computer Science: cache invalidation and naming things"

I'm very much open to suggestions for a better name.